Refactor database and visualization modules
- Removed unused imports and functions from database initialization and queries.
- Updated the elemental balance evolution query to calculate cumulative sums.
- Enhanced figure saving functionality with consistent padding and improved logging.
- Changed chart titles and sizes to better reflect the new configurations.
- Replaced area chart plotting with line chart plotting for elemental balance evolution.
- Updated tests to reflect changes in data fetching and added new tests for color functions.
- Removed obsolete plotting base utilities and refactored bar and heatmap plotting functions.

+++ hsrws/visual/config/config_chart_sizes.py
@@ -44,7 +44,7 @@ def get_default_font_scale_by_chart() -> Dict[str, Dict[str, float]]:
     }
 
 
-def get_default_chart_sizes() -> Dict[str, Tuple[int, int]]:
+def get_default_chart_sizes() -> Dict[str, Tuple[float, float]]:
     """
     Get default chart sizes optimized for Instagram.
